[refactor] 파일 위치 및 이름 변경
